+++ cov_tree/__init__.py
@@ -1,5 +1,5 @@
 from .version import __version__
-from .node import Path, CovStats
+from .node import Path, CovNode, CovModule, CovFile
 from .builder import build_cov_tree
-from .algo import max_depth, iterate, follow_path
+from .algo import max_depth, iterate
 from .print import print_tree, cov_color, get_available_tree_sets