+++ cov_tree/print.py
@@ -1,7 +1,7 @@
 from typing import Callable, Sequence
 from termcolor import cprint
 
-from .node import CovStats
+from .node import CovNode
 
 
 _TREE_SET = {
@@ -17,26 +17,27 @@ def get_available_tree_sets() -> list[str]:
 
 def cov_color(
         coverage: float,
-        thresholds: tuple[float, float, float] = (0.8, 0.8, 0.95),
+        thresholds: tuple[float, float, float] = (0.5, 0.5, 0.95),
 ) -> str | None:
     if coverage >= thresholds[2]:
-        return 'green'
+        return 'light_green'
     elif coverage >= thresholds[1]:
         return 'yellow'
     elif coverage >= thresholds[0]:
         return None
     else:
-        return 'red'
+        return 'light_red'
 
 
 def _print_tree(
-        node: CovStats,
+        node: CovNode,
         level_last: tuple[bool, ...],
         tree_width: int,
+        show_missing: bool,
         show_module_stats: bool,
         cov_color: Callable[[float], str | None] | None,
         tree_set: Sequence[str],
-        threshold: Callable[[CovStats], bool] | None,
+        threshold: Callable[[CovNode], bool] | None,
 ) -> None:
     is_leaf_like = (
         len(node.children) == 0 or
@@ -56,16 +57,21 @@ def _print_tree(
         end='',
     )
     cprint(
-        '{:{w}}  '.format(node.name, w=tree_width-len(tree)),
+        '{:{w}}'.format(node.name, w=tree_width-len(tree)),
         **kwargs,  # type: ignore
     )
     if is_leaf_like or show_module_stats:
         cprint(
-            '{:6,d}  {:6,d}  {:5.0%}'.format(
-                node.num_lines, node.num_missed, node.coverage(),
-            ),
+            f'  {node.num_executable_lines():6,d}'
+            f'  {node.num_missed_lines():6,d}'
+            f'  {node.coverage():5.0%}',
             **kwargs,  # type: ignore
         )
+        if show_missing:
+            cprint(
+                f'  {node.missed_lines_str()}',
+                **kwargs,  # type: ignore
+            )
     print()
 
     if not is_leaf_like:
@@ -75,6 +81,7 @@ def _print_tree(
                 child,
                 level_last=level_last+(last,),
                 tree_width=tree_width,
+                show_missing=show_missing,
                 show_module_stats=show_module_stats,
                 cov_color=cov_color,
                 tree_set=tree_set,
@@ -82,7 +89,7 @@ def _print_tree(
             )
 
 
-def _max_tree_width(tree: CovStats, tab: int = 4) -> int:
+def _max_tree_width(tree: CovNode, tab: int = 4) -> int:
     return max(
         (tab + _max_tree_width(child, tab) for child in tree.children),
         default=len(tree.name),
@@ -90,22 +97,27 @@ def _max_tree_width(tree: CovStats, tab: int = 4) -> int:
 
 
 def print_tree(
-        tree: CovStats,
+        tree: CovNode,
+        show_missing: bool = False,
         show_module_stats: bool = True,
         cov_color: Callable[[float], str | None] | None = None,
         tree_set: str = 'fancy',
-        threshold: Callable[[CovStats], bool] | None = None,
+        threshold: Callable[[CovNode], bool] | None = None,
 ) -> None:
     tree_set_ = _TREE_SET[tree_set]
     tab = len(tree_set_[0])
     tree_width = _max_tree_width(tree, tab)
 
     cprint('{:{w}}  {:>6s}  {:>6s}  {:>5s}'.format(
-        '', 'Stmts', 'Miss', 'Cover', w=tree_width), attrs=['bold'])
+        '', 'Stmts', 'Miss', 'Cover', w=tree_width), attrs=['bold'], end='')
+    if show_missing:
+        cprint('  Missing', attrs=['bold'], end='')
+    print()
     print()
     _print_tree(
         tree, tuple(),
         tree_width=tree_width,
+        show_missing=show_missing,
         show_module_stats=show_module_stats,
         cov_color=cov_color,
         tree_set=tree_set_,