+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Change Log
 
-# Unpublished
+## 0.2.0
+
+* entire redesign of node class
+* align coverage calculation with skipped lines with `coverage`
+* option to show missing lines for cmd line tool